requirements.txt
streamlit==1.28.0
pandas==2.2.0
numpy==1.26.0
plotly==5.17.0
openai==1.3.0
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
python-dotenv==0.19.1